Remove scrapet.rb
